dev/tests/acceptance/pre-install.php refactoring.
 - Move classes to `pre-install' directory.
 - Inject `CliColors` class from `PreInstallCheck` constructor.
 - Move all code from constructor in `PreInstallCheck` class to new public method `run()`.
 - Made all methods `private` (excepts public `run()`).
